requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
sentence-transformers
faiss-cpu
PyMuPDF
requests
python-multipart
